+++ src/services/ValidacaoComentarioServices.js
@@ -1,0 +1,7 @@
+class ValidacaoComentarioServices {
+    static ValidaConteudoComentario(conteudo) {
+        return conteudo.length >= 1 && conteudo.length <= 560
+    }
+}
+
+export default ValidacaoComentarioServices
