+++ src/services/ValidacaoPostServices.js
@@ -1,0 +1,7 @@
+class ValidacaoPostServices {
+    static ValidaConteudoPost(conteudo) {
+        return conteudo.length >= 1 && conteudo.length <= 560
+    }
+}
+
+export default ValidacaoPostServices
